pme: Implement an amplifier HAL
The amplifier correctly starts up the hardware and configures
it with the Music profile and then turns the amp on and off as
it is needed.

I assume that we will need to switch profiles based on the audio
use-case including likely to use the Ringtone profile for
notifications and one or more voice profiles for phone calls.

Change-Id: Ib0f1d348a83761c3f452980e0d5df26a9e45e8f2