Update patch set 1
Patch Set 1:

All my Exynos devices and hima is going to switch too (see the commit for it).
The current layout is an absolute mess.
As a conparison, Nexus devices even have ramdisk scripts in their root tree, we don't do that either (but rather keep them organized in rootdir).

Patch-set: 1