Update patch set 1
Patch Set 1: Code-Review-1

While your organization may be more sensible, sensible does not beat practical imo.

The only "sister" device in terms of hardware that we have is the oneplus3.  It follows the standard config and therefore will be much simpler to diff against if we need to to compare what we have.

AFAIK, all the HTC devices are following the existing conventions as well.

I killed Sultan's attempt to make rootdir more sensible (wtf is everything that goes in root in rootdir/etc ???) I would also say that this just makes life more painful for little practical benefit and therefore don't think it's really a good idea.

Patch-set: 1
Label: Code-Review=-1