Update patch set 2
Patch Set 2:

Setting the persist property seems very yucky to me.  It is not very nice because we are assuming that no other HTC blob or firmware is using this property and that it is only being used by libhtcril_db to set a bad cnv value when decrypted.

The change does seem like this would be (currently) safe to avoid the code in libhtcril_db.so but it also seems silly given how simple that library is.  I think we should just rewrite the library and then not risk being dependent on nothing else accessing this property.

Patch-set: 2