Update patch set 1
Patch Set 1: Code-Review-2

(1 comment)

Actually this code does have an effect in that it calibrates the amp.

The stock code does this specifically using profile 0 to cause the calibration to occur.

I can't say for sure that it's necessary but they do that on HAL initialization (as do all the other (4?) amplifiers that I've looked at) and don't wait to cause the calibration to occur at some later point in time.

I don't see what benefit there is in waiting for some random profile to get loaded to cause the calibration to occur.

Furthermore, what happens with the existing code is that the amp goes through these states:

(uninitialized) -> profile 11 -> profile 0

before switching to any other profile.  This may be important and it could be the case that you get weird behaviour if you skipped loading profile 0 (for example, by merging this commit and having the first sound played be a voice call which causes it to start using the amp without every loading profile 0).

Patch-set: 1
Label: Code-Review=-2