Update patch set 1
Patch Set 1:

It's not the same as line 63 after the next patch which changes "0" to be a different profile number.

Why do you want to change the way that HTC initialized the amp (trigger the calibration at startup) and instead wand to trigger it implicitly on the first attempt to play audio?

Patch-set: 1