Update patch set 1
Patch Set 1:

Can you test this (with the source built, not prebuilt, libbt-vendor.so) and let me know if it helps with your bt problems?

From a first pass at looking at libbt-vendor.so, that's the only obvious diff that I'm seeing but there could be more hidden in there.

My car works mostly fine (sometimes slow to reconnect) and my earbuds work perfectly on bt so it's hard for me to test this myself.

Patch-set: 1