Update patch set 1
Patch Set 1:

> Can you test this (with the source built, not prebuilt,
 > libbt-vendor.so) and let me know if it helps with your bt problems?
 > 
 > From a first pass at looking at libbt-vendor.so, that's the only
 > obvious diff that I'm seeing but there could be more hidden in
 > there.
 > 
 > My car works mostly fine (sometimes slow to reconnect) and my
 > earbuds work perfectly on bt so it's hard for me to test this
 > myself.

Yes of course, building at now. I come back with result and let you know.

The main problem is, that if i try pair with my N5 or my laptop or send a file than its turning on-off all the time and its impossible.

And working with my car too, but sometime its not connect and should turning on-off or very slow connecting compared with stock, if its connected than its ok maybe but not a best.

Patch-set: 1