Update patch set 1
Patch Set 1: Code-Review-1

> Can you test this (with the source built, not prebuilt,
 > libbt-vendor.so) and let me know if it helps with your bt problems?
 > 
 > From a first pass at looking at libbt-vendor.so, that's the only
 > obvious diff that I'm seeing but there could be more hidden in
 > there.
 > 
 > My car works mostly fine (sometimes slow to reconnect) and my
 > earbuds work perfectly on bt so it's hard for me to test this
 > myself.

Unfortunately, the error/issue still exists. When I try to pair my Nexus 5 devices subject to the bluetooth function is switched off-on and can not be realized/paired. So useless in this regard at the moment

Another question is that BLUETOOTH_QTI_SW option in bdroid_buildcfg.h is not for qualcomm chip?

Patch-set: 1