Update patch set 2
Patch Set 2:

Try the new patch set (for the original bt problems).

Something (for me) is still funky with connecting to my laptop but I don't think it's related to this low a level of config because if I connect and start the transfer fast enough it works well enough to send a 10MB file from my laptop to my phone.

I haven't tested bt audio for very long but it seems to reconnect immediately to my headset and I haven't had any audio problems in the bit of testing I did.

Patch-set: 2