Update patch set 1
Patch Set 1:

fwiw, to get the acdb ids that are in this file, I installed the stock software and then build:

https://review.cyanogenmod.org/#/c/163516/

I then renamed libacdbloader.so to "libacdbloader_vendor.so" and then installed the wrapper.  This let me grep the logcat for various use-cases to see what acdb calls were geing made to make sure we had the same ids and nothing else weird happening.

Patch-set: 1