Update patch set 1
Patch Set 1:

It doesn't fix anything on our device, op3 had an issue where flashlight tile was broken on encrypted devices. I merged this patch for testing when I was still using the m camera blobs and left it in because it is set to late_start on stock and CAF as well.

Patch-set: 1