Update patch set 2
Patch Set 2:

That's strange, I've been running this patch for a while and never experienced the black screen but I saw it in the list of bugs. I left the patch out when I built on top of your new changes and that's when I saw it first, just tested again with the patch and no black screen on 3 reboots. I'm running 1.95.91.3 firmware and I'm testing new cne, gps and perf blobs but I only experience the issue without this patch. I'll do some more testing.

Patch-set: 2