Update patch set 1
Patch Set 1: Code-Review-1

This is defining a backend.  A backend is appended to the paths (e.g. low-latency for backend speaker selects the mixer path "low-latency speaker").

There is no sign of a backend speaker-and-headphones but lots of evidence of speaker-headset in the stock mixer_paths_tasha.xml.

Is there a usecase that is broken you are trying to fix that I can reproduce?

Patch-set: 1