Update patch set 1
Patch Set 1: Code-Review-2

I'm pretty sure that this isn't necessary and doesn't actually do anything because I'd be amazed if the amp is being used at all here.

I added this commit just in case because the stock tinymix values that I captured showed this profile as being active.

Patch-set: 1