Update patch set 1
Patch Set 1:

Note: This doesn't actually fix bluetooth calling.

Prior to this change making a call while on a bluetooth headset caused the dialer to hang without successfully completing the call.

After this change there is no audio through the headset but the call completes and you can switch to speaker or earpiece to actually talk to someone.

Patch-set: 1