Update patch set 1
Patch Set 1:

In general, yes, please do send a pull request with a commit for each change that you're making to proprietary.  It makes it much easier for me to deal with.

With this set of changes, I just pulled the changes from your vendor repo myself and sorted them out so we're good here.

Patch-set: 1