Update patch set 1
Patch Set 1:

This is all very interesting. On CM13, gps will not lock on to any satellites. I'm not sure if it's specific to my Sprint device (being CDMA) or all devices. I was going to try and cherry pick from this, but it's dependent on the new N blobs is it not?

Patch-set: 1