Update patch set 2
Patch Set 2:

Do you want to give this version a try?

If it works correctly to link to the right file and all that, you can completely replace gps.conf.sprint with the stock version if you want (or keep this version if you think it is better like this).  Either way it won't affect any device other than Sprint devices.

Patch-set: 2
Label: Verified=0