Update patch set 4
Patch Set 4:

I don't suppose anyone reading this can provide logs that indicate that this might actually work on pme?  I only have 1 device that claims to support aptx and when I play audio via it, there is nothing in the logcat that matches grep -i aptx so I have no reason to believe that this actually works for us.

Patch-set: 4