Update patch set 1
Patch Set 1:

This was motivated by looking at IMS but at the same time independent of it.

That said, the current version doesn't work at all so it's not worth testing.  When I push up something not stating "WIP" it is worth testing on its own and/or with the IMS changes.

Patch-set: 1
Label: Verified=0