Update patch set 1
Patch Set 1:

"Browser" used to be included in the common config and we overrode that to provide a more feature rich browser.  I'm not sure that the Jelly change changes the fact that we want the more advanced browser.

Given that people are working on updating Gello right now, I'm not sure it's fair to consider it dead and gone...

Patch-set: 1
Label: Code-Review=0