Update patch set 2
Patch Set 2:

OEMs/ODMs can choose to enable this in the camera stack.
The SDM gralloc has special cases to handle this compression type.
I'm not sure if this is used on pme, but this is a standard 8996 property and should be enabled on all devices iirc (check Marlin, OP3, ...).

Patch-set: 2