Update patch set 2
Patch Set 2:

How are you testing that this works and that it didn't work before this commit?

I just took a build from July 1st and plugged in a USB-OTG cable and a USB thumb drive and it reads it just fine (in Settings >> Storage  and the file manager app).

Patch-set: 2