Update patch set 4
Patch Set 4:

I think what Tab was saying was that we were successfully using the same blob pulled from the unlocked variant.

At some point HTC released a new blob + modem firmware for the unlocked variant.  The new blob (for which there isn't the corresponding firmware for the CDMA variants) broke without the updated firmware.

Therefore we rolled back to the old version of the unlocked blob for the cdma devices.

Patch-set: 4