Update patch set 1
Patch Set 1:

> Build breakage is a good way to encourage maintainers to test and
 > merge it themselves...

Agreed, however Ethan left a +1 on the hiae change when he was using the same keyhandler as pme, so it was working.
On a separate note, there was no indication from you that you received this patch, or were planning on looking at. I understand you're busy, however could you maybe leave a note of acknowledgement in the comments next time, please?

Patch-set: 1